Nutrition recommendation system/requirements.txt
# Core Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
sqlite3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# Data Processing & ML
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
scipy==1.11.4

# Nutrition & Health Libraries
nutritionix==0.2.0
usda==1.0.0

# API & HTTP
requests==2.31.0
httpx==0.25.2

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
rich==13.7.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Jupyter Notebook Support
notebook==7.0.6
jupyter==1.0.0
ipykernel==6.27.1
